+++ src/ui/fileView/fileView.h
@@ -23,7 +23,7 @@ class FileView final : public ListView
      * @param Items items that shall be shown in the listview
      * @param page page that is shown, default is 1
      */
-    FileView(const irect &p_contentRect, const std::vector<FileItem> &p_files, uint8_t page = 1);
+    FileView(const irect &p_contentRect, const std::vector<FileItem> &p_files, int page = 1);
 
     FileItem &getCurrentEntry()
     {
